+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 **THIS SOFTWARE IS STILL IN ALPHA AND THERE ARE NO GUARANTEES REGARDING API STABILITY YET.**
 
 ## [Unreleased]
-Nothing so far
+- Make Auth.Grant(…) idempotent and do not unnecessarily add smaller scopes
+- Support extending permissions via Auth.Grant(…)
+- Add boolean return value to Auth.Grant(…) to indicate if a new permission was granted
+- Add Auth.Revoke(…) to remove permissions
+- Fix flaky unit test TestBrain_Memory
+- Fix flaky TestCLIAdapter_Register test
 
 ## [v0.7.0] - 2019-04-18
 - Add ReceiveMessageEvent.Data field to allow using the underlying message type of the adapters